chore: remove old accounts page
